chore: dashboard updates
