+++ apps/evm/src/pages/Fusion/components/Dashboard/Dashboard.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 
 import { useGetUser } from '../../../../hooks';
 import { LoginSignUp } from '../LoginSignUp';
-import { StyledCard } from '../PartnersSection/PartnerCard.style';
+import { StyledCard } from '../ChallengesAndEcosystem/ChallengesAndEcosystem.style';
 import { ProjectStatus } from '../ProjectStatus';
 import { UserStats } from '../UserStats';
