+++ apps/evm/src/pages/Fusion/components/index.tsx
@@ -1,9 +1,8 @@
-export * from './Challenges';
 export * from './Dashboard';
 export * from './Info';
 export * from './Leaderboard';
 export * from './LoginSignUp';
-export * from './PartnersSection';
+export * from './ChallengesAndEcosystem';
 export * from './ProjectStatus';
 export * from './UserStats';
 export * from './WithdrawModal';
